requirements.txt
networkx==2.8
bs4==0.0.1
graphviz==0.20
scikit-image==0.19.2
psutil==5.9.0
mmh3==3.0.0
pyhash==0.9.3
imbalanced-learn==0.9.0
Pillow==9.1.0